mpcL

+++ imitation_extended_mpc.py
@@ -289,7 +289,7 @@ for ep in range(epochs):
         #    npat = ep + 2
         
         #x0 = utils_new.sample_init_traj_dist(BS, true_dx, x_star, npat).float()
-        x0 = utils_new.sample_init_traj_dist(BS, true_dx, np.transpose(x_manual_full_H), npat)
+        x0 = utils_new.sample_init_traj_dist(BS, true_dx, np.transpose(x_manual_full_H), npat).float()
         #x0_3 = utils_new.sample_init(BS//3, true_dx)
 
         #x0 = torch.vstack((x0_1, x0_2, x0_3))
